[ALPS02779131] PMIC: fix build error
[Detail]
1. fix build error of pmic_scp_set_vcore()

MTK-Commit-Id: 415108b80adcd1bfdd7960daaf5c4eb0b41dcc00

CR-Id: ALPS02779131
Feature: Power Management

+++ drivers/misc/mediatek/pmic/mt6335/pmic.c
@@ -321,7 +321,7 @@ unsigned int pmic_scp_set_vcore(unsigned int voltage)
 }
 
 /* SCP set VSRAM_VCORE voltage, return 0 if success, otherwise return set voltage(uV) */
-unsigned short pmic_scp_set_vsram_vcore(unsigned int voltage)
+unsigned int pmic_scp_set_vsram_vcore(unsigned int voltage)
 {
 	const char *name = "VSRAM_VCORE_SSHUB";
 	unsigned int max_uV = 1200000;
